Implement @redcatbear's review findings
